rpc: handle waas api interaction in RPC instead of provider
